requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq>=1.0.0
pypdf>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# RAG Pipeline
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
qdrant-client>=1.12.0
